financial statment

+++ python_introduction/finance_calculator.py
@@ -3,6 +3,6 @@ expenses = int(input("Enter your total monthly expenses: "))
 interest = 0.05
 
 saving = income - expenses
-projected = saving * 12 + (saving * 12 * interest)
+projected = int(saving * 12 + (saving * 12 * interest))
 print("Your monthly savings are", saving)
 print("Projected saving after one year, with interest, is: ", projected)
